Configurator/Backend/requirements.txt
astroid==2.4.2
importlib-metadata==3.1.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pyserial==3.5
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.4.0
